不要な変数、コメントアウトの削除、関数の削除 privateの追加、条件式の書き換え viewmodelのフォルダ変更
