+++ app/src/main/AndroidManifest.xml
@@ -24,7 +24,7 @@
             android:name="com.google.android.gms.version"
             android:value="@integer/google_play_services_version" />
         <activity
-            android:name=".activities.MainActivity"
+            android:name=".view.activities.MainActivity"
             android:exported="true"
             android:label="@string/app_name">
             <intent-filter>
