+++ app/src/main/res/layout/fragment_achieve.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragments.achieve.AchieveFragment">
+    tools:context=".view.fragments.achieve.AchieveFragment">
 
     <ScrollView
         android:id="@+id/scrollView"
