+++ app/src/main/res/layout/fragment_notifications.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragments.notifications.NotificationsFragment">
+    tools:context=".view.fragments.notifications.NotificationsFragment">
 
     <TextView
         android:id="@+id/text_notifications"
